+++ src/aux-server/aux-web/shared/vue-components/MonacoTagEditor/MonacoTagEditor.vue
@@ -28,18 +28,6 @@
                     <span class="at-symbol">@</span>
                 </md-button>
             </div>
-            <div class="editor-errors">
-                <div v-if="scriptErrors.length === 0">No Errors</div>
-                <div
-                    v-else
-                    class="editor-button"
-                    :class="{ active: showErrors }"
-                    @click="toggleErrors()"
-                >
-                    <md-icon class="error-icon">error</md-icon>
-                    {{ errorsLabel }}
-                </div>
-            </div>
             <div v-if="signed" class="editor-signed">
                 <div>
                     <md-icon>verified_user</md-icon>
@@ -48,19 +36,6 @@
             </div>
         </div>
         <div class="code-editor-wrapper">
-            <div v-if="showErrors" class="errors-wrapper">
-                <ul class="error-list">
-                    <li v-for="(error, index) in scriptErrors" :key="index" class="error">
-                        <div class="error-count">
-                            <span>{{ error.count }}</span>
-                        </div>
-                        <div class="error-details">
-                            <div class="error-name">{{ error.name }}</div>
-                            <span class="error-message">{{ error.message }}</span>
-                        </div>
-                    </li>
-                </ul>
-            </div>
             <monaco-editor
                 ref="editor"
                 @focus="editorFocused"